+++ packages/@reflex/beta/study/01-basics.md
@@ -1,0 +1,177 @@
+# 01. База: що взагалі таке reactive runtime
+
+## 1. Що таке реактивність
+
+Реактивність - це спосіб автоматично підтримувати залежності між значеннями.
+
+Простий приклад:
+
+```ts
+const price = signal(100);
+const tax = signal(20);
+const total = computed(() => price() + tax());
+```
+
+Тут:
+
+- `price` і `tax` - джерела
+- `total` - похідне значення
+
+Якщо змінюється `price`, значення `total` має стати новим.
+
+Головне питання будь-якого runtime:
+
+коли саме перераховувати `total`?
+
+## 2. Три базові ролі
+
+У `@reflex/beta` є три ролі:
+
+- `signal`: зберігає вихідне значення
+- `computed`: обчислює похідне значення
+- `effect`: запускає побічний код, що залежить від reactive graph
+
+Важливо:
+
+- `signal` - це джерело істини
+- `computed` - це кешована функція
+- `effect` - це не значення, а реакція
+
+## 3. Push vs Pull
+
+Є два класичні підходи.
+
+### Push
+
+Коли source змінюється, runtime одразу проштовхує оновлення графом вниз.
+
+Плюси:
+
+- downstream часто вже готовий до читання
+- зручно для eager-моделей
+
+Мінуси:
+
+- запис стає дорогим
+- легко перерахувати багато зайвого
+
+### Pull
+
+Коли source змінюється, runtime не перераховує все одразу.
+Він лише позначає downstream як підозрілий.
+А реальний перерахунок відбувається тоді, коли хтось читає значення.
+
+Плюси:
+
+- запис дешевий
+- часткові читання вигідні
+- глибокі ланцюжки читаються ефективно
+
+Мінуси:
+
+- фінальне читання може саме потягнути велику перевірку графа
+- high fan-in сценарії можуть бути слабшими за eager runtime
+
+`@reflex/beta` обирає lazy pull.
+
+## 4. Що означає lazy pull у beta
+
+Коли ви робите:
+
+```ts
+setCount(10);
+```
+
+beta не йде миттєво перераховувати всі `computed`.
+
+Вона робить лише:
+
+1. змінює payload source
+2. зсуває epoch зміни
+3. маркує downstream як `Invalid`
+
+А далі:
+
+```ts
+view();
+```
+
+і лише в цей момент runtime перевіряє:
+
+- чи справді dependency змінилися по суті
+- чи треба реально перераховувати поточний вузол
+- чи можна пропустити recompute
+
+## 5. Що таке graph
+
+Усередині runtime все зберігається як граф залежностей.
+
+Якщо:
+
+```ts
+const a = signal(1);
+const b = computed(() => a() + 1);
+const c = computed(() => b() * 2);
+```
+
+то залежності такі:
+
+```text
+a -> b -> c
+```
+
+Це означає:
+
+- `b` залежить від `a`
+- `c` залежить від `b`
+
+У beta ребро читається так:
+
+`to depends on from`
+
+Тобто якщо є edge `from=a`, `to=b`, значить `b` читає `a`.
+
+## 6. Що runtime має гарантувати
+
+Будь-який пристойний reactive runtime має робити хоча б це:
+
+- не допускати "глітчів" під час читання
+- не перераховувати без причини
+- не забувати старі/нові залежності під час branch switching
+- не ламати кеш після same-value результату
+
+Для beta це виражається через версії, dirty hints і tracking epochs.
+
+## 7. Чим beta відрізняється за характером
+
+Це не "UI framework runtime".
+Це маленьке ядро reactive graph.
+
+Його пріоритети:
+
+- прості інваріанти
+- хороша продуктивність у lazy workload
+- мінімалізм
+- можливість розширити поверх ядра eager/flush шар
+
+Не його мета:
+
+- одразу розв'язувати всі сценарії
+- бути максимально feature-rich
+- приховувати архітектурні компроміси
+
+## 8. Найкорисніша mental model
+
+Думайте так:
+
+- `signal` - комірка з даними
+- `computed` - кеш функції + список прочитаних dependencies
+- `effect` - обчислення, значення якого не потрібне, але важливо виконати код
+- graph - це карта "хто кого читав"
+
+І ще:
+
+- write робить graph dirty
+- read робить graph fresh
+
+Це і є фундамент beta.
